+++ src/c/modules/bluetooth_m.c
@@ -66,6 +66,9 @@ static void prv_get_weather_error_symbol(char *err_symbol) {
     case WEATHER_LOCATION_ERROR:
       snprintf(err_symbol, 2, "%s", "E\0");
       break;
+    case WEATHER_API_BANNED:
+      snprintf(err_symbol, 2, "%s", "X\0");
+      break;
     default:
       snprintf(err_symbol, 2, "%s", " \0");
       break;
@@ -109,10 +112,7 @@ static void prv_populate_bt_layer(Layer *me, GContext *ctx) {
   }
   //if (settings_get_ClockFormatSettings() == CF_RESPECT)
   if (strcmp(settings_get_ClockFormat(), "%I:%M") == 0) {    
-    struct tm *tick_time;
-    // time_t temp = time(NULL);
-    tick_time = get_Time();
-    graphics_draw_text(ctx, tick_time->tm_hour < 12 ? "AM" : "PM" , \
+    graphics_draw_text(ctx, get_Time()->tm_hour < 12 ? "AM" : "PM" , \
     fonts_get_system_font(FONT_KEY_GOTHIC_18), \
     GRect (60, 0, 20, 20), \
     GTextOverflowModeWordWrap, \